Use a FlinkUserClassLoader if no context cl set
This commit sets the FlinkUserClassLoader as the Thread's context
class loader if such a class loader was not previously set. When
statefun is started by Flink a context class loader is already set
(and it is of type URLClassLoader) but when statefun is started differently
(for example JUnit) the context class loader might be of a different type.
This commit unifies the context class loader type.